Use zlib-ng to dramatically improve compression performance
After doing more research on compression algorithms and finding out
just how similar Yaz0 and LZ77-style compression algorithms are,
I tried to implement some common tricks to help improve the extremely
poor compression performance due to the sliding window search.

That led to the hash-table optimisation, which is used in several
LZ77 encoders. But the implementation was still extremely slow
compared to gzip. And probably badly implemented.

It turns out that many more tricks were needed for fast compression.

After stumbling upon zlib-ng and seeing how well it is optimised
and all the intrinsics I decided it was best not to reinvent the wheel.

This adds a copy of zlib-ng with the following modifications:
* The window size is reduced to 4K to match Yaz0.
* The compress function and the stream structures were changed to take
  a callback that is invoked every time a distance/length pair or a
  literal is emitted. (I'm not proud, but it works.)
* MAX_MATCH was NOT increased. zlib assumes it is equal to 258 in
  too many places and an increase resulted in worse compression ratios.

The compression now relies entirely on zlib for the heavy lifting
and is dramatically faster as a result.

Useful resources:
* K. Sadakane & H. Imai (2000), "Improving the Speed of LZ77 Compression by Hashing and Suffix Sorting". https://pdfs.semanticscholar.org/956d/e955607e92e33151db4216b93e51226f2b87.pdf
* E. Chen (2019), "Understanding zlib". https://www.euccas.me/zlib/